LTG-315: Don't expose NotifyRequest to API
The `NotifyRequest` object has the concept of `destination` which can be either an email address or a telephone number. This won't be known by the frontend. The frontend will know the session ID (which it provides as a header) and the e-mail address, which is used to validate against the session ID provided. The actual email address and/or phone number for notifications will be pulled from the user storage before passing onto SQS to deliver to Notify.