BAU: Return optional UserProfile from AuthenticationService
First part of an incremental refactoring to make the return type of AuthenticationService#getUserProfileByEmail an Optional rather than returning nulls.  At this stage we are only refactoring cases where nullability was already being taken into account.

The original version of the method ought to be deprecated, but as we are no longer allowed to have unresolved warnings, marking it as deprecated is too awkward to be worthwhile. :(