AUTH-1397: use dynamodb transactions for update email
Use transactions to maintain database integrity given that deletes and creates are required in two separate tables for the operation to succeed.

DynamoDB mappers only support transactions on a single table if table overrides are in use.  This means using a different api which requires converting fields to AttributeValue in order to use transactions.