BAU: Consent to share off in the acceptance test client
Consent to share is still switched-on in the acceptance test client because there were existing tests that had been built to use it while it was still required.  The decision has been made the consent will be off for all clients so it makes more sense to remove these steps from the tests in order to reflect what is live.