AUT-590: support ui_locales in /authorize
If either of the supported languages en or cy are passed in ui_locales then the language cookie will be set to switch to the correct language.  If the paramater is missing, or no supported languages are provided then the cookie will not be set, and frontend user choices will continue to be respected.