AUT-1441: Add service for Bulk Email Users
Adds the basics for accessing and updating information in the bulk email users table, introduced temporarily to support sending a bulk email to update users on terms and conditions. Also adds test extension and integration test.

The Base Dynamo Service has also been slightly modified to expose the non-enhanced client. This should not affect existing code, but means that we can run non-enhanced queries (queries that do not return the domain objects) since there are some known oddities with limiting the number of results with enhanced queries.

Co-authored-by: Daniel Besbrode <daniel.besbrode@digital.cabinet-office.gov.uk>