AUT-1443: bulk email sending lambda and tests
Lambda triggered by a cloudwatch scheduled event which reads the contents of the bulk-user-email table, retrieves the email corresponding to the subject id, sends an email to the user, then updates the status for the user email in the table.

The table is read and the emails are sent in batches with a configurable size.  After a batch has been processed sending can be paused (sleeped) for a configurable period in order to throttle email sending if required.

Includes unit and integration tests for the handler.

'getUserProfileFromSubject' in the DynamoService has been updated to throw an exception when a user is not found by subject id.