AUT-1546: Handle 404 responses received from DCMAW
In the case that DocAppCallbackHandler receives HTTP 404 from CRI request, return HTTP 302 response to HMRC. This redirect contains the following parameters in the URI (to match what is expected in a "user abort" scenario): error="access_denied" and error_description="Not found".

To achieve this, UnsuccessfulCredentialResponseException class was modified to hold the corresponding HTTP code.