AUT-1567: Filter users by terms and conditions versions
Some users have already accepted a version of the terms and conditions that state the new retention period. We therefore want to filter these users out from those that we are sending the bulk email to. We support passing in multiple terms and conditions versions, since there have been multiple updates with the new retention period.

Users who have not accepted any terms and conditions, where this attribute is missing, will be filtered out of the results.  This is because these users did not succeed in completing the account registration process, do not have verified accounts, and have never logged in.  They should therefore not be contacted.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>