AUT-1567: Filter audience loaded by terms and conditions versions
This allows us to filter out users who have accepted later terms and conditions versions and therefore do not need to be notified of the update.

This reads the excluded terms and conditions from an environment variable and also populates the environment variable (BULK_USER_EMAIL_EXCLUDED_TERMS_AND_CONDITIONS) via terraform.

We throw an exception when excluded terms and conditions config is missing to avoid running under a misconfigured lambda.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>