AUT-1633: Do not send email to users who have accepted excluded T&Cs
When we populate the table of users to send the bulk email to, we exclude those users on later versions of the T&Cs who have already accepted the retention period change, in order to not unecessarily email users.

However, it is possible for a user who was on an old version of the T&Cs when we populated the table, to have since accepted the updated T&Cs and so be on a different version to when we originally populated the table. Therefore, this commit adds an extra layer of filtering when we are sending the emails to ensure that we do not send emails to these users.