AUT-1633: Fail the bulk email sender when env not properly configured
Throw an error when there are no included terms and conditions set as an environment variable since this indicates misconfiguration. This can lead to no filtering occuring when sending emails (which would mean we would send an email to a user who had accepted a later version of the terms and conditions since the original table of users to send emails to had been created).