AUT-1695: Add access policies for delivery receipts lambda
We are temporarily granting access to the delivery receipts api to the following tables:
* bulk email user table - read and update access
* user profile table - read only

This will enable us to update the status of a user in the bulk email users table according to the delivery receipt status (to note, for example, cases where the delivery receipt indicates failure). We need access to the user profile table to convert an email address (received in the delivery receipt) to a subject id (used for the bulk email users table).

This access can be revoked once work to send the bulk email for terms and conditions update has been completed.