AUT-1472: Configure batch send lambda to send for next two days
This configures us to send up to 100,000 emails tomorrow and the next day.

The cron expression breaks down as follows:
* 0/06 trigger the lambda every 6 minutes
* 10-13: on the hours between 10 UTC and 13 UTC (so on the hours of 11 BST and 14 BST, which works out to be sending every 6 minutes between 11 and 3pm BST)
* 26,27 SEP ? 2023: on 26 and 27 September

This sets us up to send up to 100,000 emails, since the batches run every 6 minutes are of 2,500, so 25,000 an hour for 4 hours.

We may not hit 100,000 if a batch does not complete in the 6 minutes before the subsequent batch starts - in this case, since we are only allowing one concurrent execution of a lambda at a time, the next batch will not be triggered.