AUT-1467: Bulk Email switch for delivery receipts lambda
Adds a switch to turn bulk email on and off in the NotifyCallbackHandler.

This is so the lambda does not try to describe the user-profile table when the bulk email deployment switch is off (deploy_bulk_email_users_count).  If the deployment switch is off the NotifyCallbackHandler does not have the iam rights to access the user-profile table, so the lambda fails to start and throws errors.