AUT-1850: Add lambda terraform for placeholder API endpoint
Note that we only want to deploy this infrastructure in build. We are not, however, controlling this by counts in the terraform but will instead control this within the pipeline, to avoid having to implement a cascading set of counts that resemble a no-op terraform, which we've found to be flakey.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>