Create stub-endpoint-module
This is a variation of the existing endpoint-module, but without the
parts which modify the API Gateway. This is to be used for stubs, and
endpoints should be configured outside of the module, via an openapi
specification passed in the `body` attribute on the
`aws_api_gateway_rest_api` resource.

This allows stubs to have fine-grained control over the structure of
the api endpoints / methods.

Documentation: https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/api_gateway_rest_api#openapi-specification