ATO-175: Add AccountInterventionService
Add AccountInterventionService and associated classes.

The AccountInterventionService is a proxy wrapper to the external Account Intervention Service. Our representation makes an HTTP call to the actual service and parses the result into our internal representation. At this stage we are only interested in a subset of the returned information, namely the account status.

If the Account Intervention Service is disabled our `AccountInterventionService` will return an AccountStatus with all false values, indicating there is no intervention.

If the Account Intervention Service is enabled, but we encounter an exception at any point when connecting to the service, we catch the exception and rethrow an "AccountInterventionException"