BAU: Refactor to add new method signature for submit audit event
We have very verbose code to submit audit events. In almost all cases, we retrieve information from the userContext in the same way. Introducing a new method to allow us to pass in the user context to the audit event submitter and let it extract the relevant information makes this a bit less verbose.

Note: in some cases we were retrieving the session id from the session, not the user context. I don't think there's a good reason to do this differently in these cases (it should exist in both) therefore I've switched these to reading from the user context too.