fix shared deprecations
ref. the removal of the `acl` parameter from the `aws_s3_bucket`
resource: I attempted to replace this with an `aws_s3_bucket_acl`
resource, as this is the recommended approach. However, our buckets are
set up with the object ownership set to `BucketOwnerEnforced`, which
prevents the use of ACLs [1]. I have therefore explicitely set the
object ownership policy to `BucketOwnerEnforced`, which results in a
no-op to the existing buckets when terraform is run.

[1] https://docs.aws.amazon.com/AmazonS3/latest/userguide/about-object-ownership.html