AUT-2146: Add pending email check queue
This creates the queue, which will allow us to check email addresses that users use to create an account or update on their account. It will be written to from this aws account, and read from the email-check aws account

The queue id is added as an output of the shared terraform, since we will need this in the lambdas to write to the queue.