AUT-2485: Skip prod config by default
This change makes it so that the `set-up-sso.py` script will not add
production account profiles to the AWS config file unless the
`INCLUDE_PROD` environment variable is set to `YES I AM SURE`.

This is to prevent accidental operations on production accounts.

Additionally, use `boto3` to create the OIDC session instead of the
`aws` CLI. This is more reliable.

Split out the role and account checks into separate functions to make
them testable.