AUT-2552: Add applied at timestamp to account interventions response
This is returned to us by the account interventions api, but we do not pass it back in the api response to the frontend. Exposing this information to the frontend will enable us to add logic to ensure that people do not go through a password reset journey twice before the account interventions api has been updated with the fact that the user has reset their password, while also ensuring that newer interventions are respected.