AUT-2553: Allow AuthCodeHandler to accept password reset timestamp
We have added the ability for the auth code handler to accept a password reset timestamp from the frontend request. When it exists, it will be stored in the auth code table. It is an optional parameter.

This time stamp is to be passed to orchestration for comparison with the time the password reset AIS status was applied. If password was reset later than the AIS status was applied, a user will be allowed to continue their journey.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>