AUT-2553: Save passwordResetTime in AccessTokenStore when it exists in auth code store
Updates the token handler to pass through a password reset time from the auth code store to the access token store.

This will enable us to access this field in the user info handler, meaning we can pass the information back to orchestration. This will allow them to implement logic that ignores certain account interventions if a password has recently been reset.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>