AUT-2553: Send password reset time to orch As claim
Sends the password reset timestamp stored in the Access Token Store to orchestration via claims attached to user info. This is so that they can check this timestamp against the time the password reset intervention was set, in order so to decide whether a user can continue on their journey.

Co-authored-by: Becka Lelew <becka.lelew@digital.cabinet-office.gov.uk>