AUT-2598: Return some of a user's MFA details in the ResetPasswordRequestHandler
If a user who is locked out due to too many incorrect password attempts starts a new frontend session, we do not make them complete a 2FA step. This is because we do not set the relevant information in their frontend session to signify that they need to complete a 2FA step when they are locked out, since we return an error from their CheckUserExists response, which is where normally we'd get this information from.

By returning the MFA type directly from the reset password request handler, we can make sure that users have to go through the 2FA step when resetting their password in a new session, when they are locked out due to too many incorrect attempts.

To do this, we needed to change the status code of the reset password request from a 204 (no content) to a 200. Without this, the api gateway strips out the response body.