AUT-2598: Save email in user session if user is blocked
When a user is locked out due to too many incorrect password attempts, we return an error from the CheckUserExistsHandler before we save anything in the session. In the frontend, they then get redirected to a page where they can ask to reset their password, which then calls the reset password handler. However, if they have started a new frontend session since being blocked, their email will not exist in the session and therefore when we attempt to validate the session's stored email against the email in the reset password request body, we will return an error.

Ensuring that we save a session with an email when a user is locked out due to too many incorrect password attempts means that they will be able to reset their password in the same session.