OLH-1640: Add template ID for new confirmation email
In the Home team we've released a new feature that allows users to view
their sign in history and report anything they don't recognise to the
Fraud team. As part of this journey we send them a confirmation email
through GOV.UK Notify and we'd like to be able to report on the delivery
of these emails.

Add the production ID for this new template to the config for the Notify
callback lambda so metrics are recorded in Cloudwatch metrics and passed
through to the Grafana dashboard[1].

[1]: https://g-8e26091ad7.grafana-workspace.eu-west-2.amazonaws.com/d/ufjJdaR4z/authentication-notify-metrics?orgId=1