AUT-2022: Fix addMFAMethodV2
Dynamo seems to need a void setter in order to be able to save and retrieve an attribute in the database.  Have created an additional 'with' method to do what the setter was doing before.

Note that the logic for V2 will be different for V1, as V1 only supported a single method which could only be replaced, whereas V2 will support a proper list of methods.  This means the 'removeIf' part of the setter can be removed.