AUT-1506: Remove synthetics_users variable from integration tfvars
There is already a secret set for this, but because we set it to an empty string it overrides the variable we have for it set as a terraform variable.

This seems to be why the create account smoke test is failing, since it depends on deleting the synthetics user. At the moment, we don't set a synthetics user, meaning that the request to the test services api 404s (there is no user set in the environment variable to delete), and the smoke test fails as we attempt to create an account as a user where we have not deleted their account first.

I have also set a default on the variable in the case of it not being present in secrets.