AUT-1506: Remove synthetics_users variables from all tfvars files
This was overriding a value set in secrets, and causing the create account smoke test to fail. The test was failing as it it depends on deleting the synthetics user. At the moment, due to the overriding value being empty, we don't set a synthetics user, meaning that the request to the test services api 404s (there is no user set in the environment variable to delete). This causes the smoke tests to fail as we attempt to create an account as a user where we have not deleted their account first.

Note that this variable is set as a secret for build and production. Staging does not have the variable, but smoke tests are not run in this environment and it's OK therefore for it to default to this empty string.