AUT-2488: Create request & response for ticf cri stub
Created the request and response classes for the TICF CRI Stub, as well as added the ability to handle the request and provide a response in line with these classes in the handler.
Currently there is no logic in the handler - the request is validated to match the TICFCRIRequest class and a default response is sent back.