ATO-430: Add authorizer to orchestration frontend API Gateway
Add an authorizer to the orchestration frontend route of the di-authentication-api API Gateway. The authorizer permits all IP addresses in production and integration, but only permits specific IPs in lower environments (sandpit, build, and staging).

The authorizer is configured to cache its authorization result differently in production/integration vs lower environments. In the former, caching is based on apiId with a TTL of 1 hour, which means that the cached (permitted) result will be used for nearly all requests. This configuration means that there isn't a significant performance drop by using an authorizer in these environments.

In lower environments, where performance is not as much of an issue and authorization depends on IP address, caching is based on sourceIp with a TTL of 1 minute.