AUT-2637: add new fields for audit events
Added new fields to audit events in line with TxMAs requirements. See ticket AUT-2637 for details.
Added an Optional input parameter to the removeAccount method in the AccountDeletionService. This
is to handle the RemoveAccountHandler's implementation of removeAccount requiring the input to
calculate values for the audit event, while the ManuallyDeleteAccountHandler does not require
those values for its audit event.