ATO-430: Create OrchFrontendAuthorizerHandler
This handler is invoked when the /orch-frontend method is called on API Gateway. It returns an Allow policy in production environments or when the source IP address is in the valid range. In other cases it throws an exception and then API Gateway returns an unauthorized response.

The generated policy allows all routes (ie. the path is set to `orch-frontend/*`). This means that the cached policy will apply to all routes.

Note that AuthPolicy has been copied from account-management-api