INC-760: Do not return an mfa method when none exists
We were defaulting to returning an MFA method of SMS even when none existed. This meant that when a user with a partially created account was going through a password reset journey, they were hitting an error (because we think they have an MFA method of SMS but actually they do not have a verified phone number.