INC-760 Do not return SMS mfa type if a phone number is not verified
We were previously assuming in the mfa helper that a user had SMS mfa by default. This made sense at some point, since we were only using this method once a user had successfully logged in, so they must have had SMS. However, we're now returning this information from the check user exists endpoint, at which point a user might not have fully set up 2fa.

Basing the MFA method on whether a user has a verified phone number or not means that for users with no verified phone number, we return a none as a 2fa method. This fixes a bug in the password reset journey, where a user without a fully set up account cannot reset their password because they hit an error screen on the 2fa step. They were hitting this 2fa step since the backend was returning that they had SMS MFA setup, so we were routing them through that journey without them actually having a verified phone number.

It should be noted that the logic for this MFA helper is slightly confusing to look at at first glance, because depending on MFA type we store this information in different ways - if an Auth app, we store this information in the user credentials table, but if SMS we do not store this information in the same table, but instead base it on whether the user has a verified number, and no auth app set up. Future work on MFA method management should fix this oddity.