BAU: Use new submitAuditEvent in Login Handler
This switches the login handler to use the new version of submitAuditEvent that takes a TXMA audit user, instead of taking all the fields individually.

This makes the calls to the audit service less noisy.

Some duplicate test assertions on the audit event submitted in the various happy path cases has also been removed, as we don't need to assert on audit event submitted every time.

- Refactor one use of `submitAuditEvent` in the auth LoginHandler to use the new `submitAuditEvent` with smaller signature.
- Add equality methods to the `TxmaAuditUser`.
- Create `user` object by using the `TxmaAuditUser` builder functionality.