AUT-2640: Update audit event component id value to 'AUTH'
Security require the component id field to be populated with a signifier
that allows them to distinguish audit events from our service (or component) from
audit events from other services.

Previously we were populating this field getting the env var OIDC_API_BASE_URL. This
was not declared in our terraform and so was returning null anyway, but even if the
value for OIDC_API_BASE_URL were returned, that would not be an appropriate
componentId for auth as it is not our service uri.