AUT-2578: Retrieve X-Forwarded-For header as ip regardless of case
As part of setting common headers on all requests, we're going to start setting the forwarded for header from a common library instead of directly in our code, which uses lower casing. This commit allows case-insensitive retrieval of this header, to ensure that when we switch over our method of setting the method, we can still get the IP address from the header.