RFF-3: Introduce new method to return headers
This is a step in a refactoring to simplify how we get headers. This introduces a new method which returns an optional, and uses this in other places in the request header helper. Eventually we will replace all calls to other methods in the request header helper with this new method.