AUT-2578: Make orchestration IP header retrieval case insensitive
As part of setting common headers on all requests, auth will start setting the forwarded for header from a common library instead of directly in our code, which uses lower casing. This commit allows case-insensitive retrieval of this header to ensure that calls from the authentication frontend (for example to the auth code handler) still allow an IP address to be set on an audit event.