AUT-2663: Introduce additional parameter for audit service submission
This introduces a restricted parameter for the audit service. It is currently unused, so this is a preparatory commit to separate out making the changes required to compile the project from the behaviour. The idea is that this will eventually be used to submit all information we put in the restricted section of audit events, but currently just contains the encoded header.