AUT-2264: Add txma audit header to the UserContext to allow retrieval in the lambdas when building a txma audit event.  This commit makes use of this in the LoginHandler to establish how it will work.
Added test to confirm the device details header is correctly retrieved and added to the UserContext so that the AuditService can add it to the audit event it publishes.

Changed the test to use a representative test value for the base64 encoded device details header and improved the readability of the code that builds the test set of headers.