ATO-617: Introduce `ParsedLogoutRequest` class
The purpose of the class is for convenince - to keep track of the many variables which may or may not be present in a logout request. It handles the parsing of a logout request into an object whose fields are then used during the logout flow.
The unit tests ensure the request is correctly parsed for the different fields that may be missing from a logout request.