AUT-2802: rewrite build and deploy workflows
Previously, we had one workflow which built all the API modules. When
this completed, 9 separate workflows then started to create terraform
artifacts and push them to the pipeline-trigger (promotion) bucket.

This commit reimplements this process using a single workflow.

As gradle will now build a binary-identical artifact if the source code
has not changed, we can avoid pushing a functionally identical blob
with a different SHA to the promotion bucket by reusing the signed
artifact generated by the previous pipeline run.

The new workflow is as follows (simplified):

1. Collects the necessary secrets and environment variables
2. Runs a matrix build to build all the API modules:
   1. Build the java artifact
   2. Check if the artifact sha exists in the dynamodb table.
   3. If it does: check the signed artifact exists in the destination
      bucket, and if it does, use it. If it doesn't, act as if it was
      not found in the dynamodb table. Delete the stale reference.
   4. If it does not: sign the new artifact and fetch the result. Store
      the result in the dynamodb table.
   5. Upload the signed artifact and a file `${module-name}_commitsha.txt`
      to github's artifact store.
3. Upload the shared terraform zip to the promotion bucket
4. In parallel:
   - Build and push terraform artifacts for the simple modules (those
     that are 'just' terraform + 1 java zip) to the promotion bucket,
     retrieving the signed zip from the gha artifact store
   - Build and push the OIDC terraform artifact (which requires a few
     java zips) to the promotion bucket, retrieving the signed zips from
     the gha artifact store
5. Push a 'trigger' artifact to the promotion bucket, to signal to the
   pipeline that all the artifacts are ready to be deployed.

The 'artifact reuse' strategy uses dynamodb to store information about
signed artifacts. When a new artifact is built, the following are stored

- module_name: the name of the module
- artifact_hash: the SHA256 hash of the artifact
- signed_object_key: the S3 key of the signed artifact
- commit_sha: the SHA of the commit that generated the artifact

When a new artifact is built, the workflow will query dynamodb to see if
there is a previous artifact with the same unsigned hash. If there is,
it will use the signed artifact from that run. Otherwise, it will sign
the new artifact and store the result in dynamodb.

We do check whether the file exists in the destination bucket before
blindly using it. if it does not, we remove the stale reference before
signing the new artifact.

The `${module-name}_commitsha.txt` file is a simple text file that
contains the commit SHA of the source code that was used to build the
artifact. This can then be consumed by terraform to link the lambda
deployment back to the source code that generated it. Previously, we
only had access to the SHA of the commit that triggered the build, which
was often not actually related to the java code.

NB: this requires the 'build' GHA AWS role to have s3:GetObject
permissions on the source and signed buckets.