AUT-2802: update build.gradle to build repeatably
This should ensure that the zip files are built in a repeatable way,
assuming that the build environment is consistent, and that the build
cache is persisted between builds. Testing has shown that with these
changes, the zip files are identical between builds (ie. SHA256 hashes
are the same between builds where the source code has not changed).

This is implemented by updating the `buildZip` task for all the APIs to:
- exclude empty directories from the zip
- not preserve file timestamps in the zip
- sort the files in the zip