AUT-2264: Updated the integration tests to provide the device_information header.
Introduced a new method in AuditAssertionHelper to allow the integration tests that test handlers that do not pass on the device_information into the audit events to continue to test that the correct number of events are being sent.

For all the handlers that do pass on the header in the audit event the existing method checks that the mandatory device_information is provided in the restricted section of the audit event.