BAU: add workflow to run pre-commit on PRs
We have a pre-commit config file which specifies some checks that should
be run before a commit is allowed. Whether it is run, though, is
dependent on the developer's local setup. This workflow ensures that
those checks are run on every PR.

pre-commit is run on against all the commits between the base and head,
which ensures that only and ALL the files modified in the PR are
checked.

Additionally, add some extra checks to the pre-commit config file:
- prettier (mostly for yaml, but will also check js, json, etc)
- python linters (pylint, flake8, black, isort)
- shellcheck (checks shell scripts)
- actionlint (checks the syntax of github actions)