AUT-2802: Rewrite dev deploy using new workflows
Rewrite the dev deploy workflow to use the reusable workflows. This
simplifies the workflow and makes it easier to maintain - changes made
in the reusable workflows will be reflected in all workflows that use
them.

`deploy-api-modules-dev.yml` and `deploy-api-modules.yml` should only
differ in the `on` trigger and the `set-up` job. All the other jobs
should be the same in each, to ensure that dev builds are as prod-like
as possible.