BAU: Upgrade all terraform lock files + checks
Versions had been changed within terraform, but the locks hadn't been
upgraded.

Added a pre-commit hook to ensure that the lock files are regenerated
in future when the main `versions.tf` is updated.

Additionally, added `destroyed-symlinks` and `check-symlinks` to the
pre-commit hooks to ensure that symlinks are not broken, which would
detach individual modules' versions from that of the main `versions.tf`.

Finally, add a few more pre-commit hooks to ensure that the terraform
files are linted and formatted correctly.