AUT-2779: Only deploy SQS trigger in supported envs
Added `support_email_check_enabled` to the utils module so we only deploy the email writer result SQS event trigger in supported environments. Set those environments to sandpit, build and staging.

This is to avoid cross account confusion and user journey breakages when a user creates an account in sandpit, the risk assessment is received and the writer lambda in authdev* or dev writes to it's own DynamoDB table