AUT-2779: Reconcile email check result message properties
The authentication-check-experian repo was writing messages with a different structure as to what authentication-api was expecting to receive. Here we make it match what is received.

There could be further opportunity to change the names of the DynamoDB fields, but that goes beyond the scope of AUT-2779 and is functionally irrelevant at the moment.