BAU: Enable identity in the build environment
To facilitate testing of identity journeys in auth, where there are still identity specific pathways.
Makes the build environment match production.
As there is no ipv stub in build it will not be possible to complete an identity journey back through orchestration, but it will still be possible to test the entire auth journey.